+++ CHANGELOG.md
@@ -12,10 +12,12 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added multiple routing keys to a single queue for monitoring.
 
 ### Added
+- \_convert_data:  Private function to convert data.
 - \_process_queue:  Private function to process message queue.
 
 ### Changed
-- process_msg:  Replaced a section of code with call to \_process_queue function.
+- process_msg:  Replaced a section of code with call to \_convert_data function.
+- \_convert_data:  Replaced a section of code with call to \_process_queue function.
 - monitor_queue:  Refactored function to initialize and monitor multiple queues.
 - callback:  Added routing key information to log entries.
 - process_msg:  Refactored function to handle multiple queues and also multiple data types in the message body.